requirements.txt
streamlit
qrcode
opencv-python-headless
Pillow